backend/Data_Analysis/requirements.txt
functions-framework==3.*
google-cloud-language==2.13.4
